feat: added system as it's own table so cascading behaviour will work on deletion
Also refactored table ids and stores to use int64.

+++ internal/models/print_job.go
@@ -5,8 +5,8 @@
 package models
 
 type PrintJob struct {
-	Id           int     `json:"id"`
-	DocumentId   int     `json:"document_id"`
+	Id           int64   `json:"id"`
+	DocumentId   int64   `json:"document_id"`
 	CupsJobId    *string `json:"cups_job_id"`
 	Status       string  `json:"status"` // pending, printing, completed, failed
 	SubmittedAt  int64   `json:"submitted_at"`
